Refactor class models for optimize, create DataLoader class

+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/AgeRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AgeRepository extends CrudRepository<Age,Integer> {
+public interface AgeRepository extends CrudRepository<Age,Long> {
 }
