+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/AnomalyRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AnomalyRepository extends CrudRepository<Anomaly,Integer> {
+public interface AnomalyRepository extends CrudRepository<Anomaly,Long> {
 }
