+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/AssetRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AssetRepository extends CrudRepository<Asset,Integer> {
+public interface AssetRepository extends CrudRepository<Asset,Long> {
 }
