+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/CharacterRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CharacterRepository extends CrudRepository<Character, Integer> {
+public interface CharacterRepository extends CrudRepository<Character, Long> {
 }
