+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/ExperienceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ExperienceRepository extends CrudRepository<Experience, Integer> {
+public interface ExperienceRepository extends CrudRepository<Experience, Long> {
 }
