+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/LanguageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LanguageRepository extends CrudRepository<Language, Integer> {
+public interface LanguageRepository extends CrudRepository<Language, Long> {
 }
