+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/RankRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RankRepository extends CrudRepository<Rank, Integer> {
+public interface RankRepository extends CrudRepository<Rank, Long> {
 }
