+++ src/main/java/com/zyknafein/roleplaycharactereditor/repository/StatsPointRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StatsPointRepository extends CrudRepository<StatsPoint, Integer> {
+public interface StatsPointRepository extends CrudRepository<StatsPoint, Long> {
 }
